projects/web_api/requirements.txt
absl-py==2.1.0
accelerate==1.4.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
albucore==0.0.23
albumentations==1.4.24
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
astor==0.8.1
async-timeout==5.0.1
asyncio==3.4.3
attrdict==2.0.1
attrs==25.1.0
babel==2.17.0
bce-python-sdk==0.9.29
beautifulsoup4==4.13.3
black==25.1.0
blinker==1.9.0
boto3==1.37.8
botocore==1.37.8
braceexpand==0.1.7
brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.1
cryptography==44.0.2
cssselect==1.2.0
cssutils==2.11.1
cycler==0.12.1
cython==3.0.12
datasets==2.14.4
decorator==5.2.1
dill==0.3.7
distro==1.9.0
doclayout-yolo==0.0.2b1
einops==0.8.1
et-xmlfile==2.0.0
evaluate==0.4.3
exceptiongroup==1.2.2
fairscale==0.4.13
fast-langdetect==0.2.5
fastapi==0.103.0
fasttext-predict==0.9.2.4
filelock==3.17.0
filetype==1.2.0
fire==0.7.0
flask==3.1.0
flask-babel==4.0.0
flatbuffers==25.2.10
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
grpcio==1.70.0
h11==0.14.0
h5py==3.13.0
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.2
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imgaug==0.4.0
iopath==0.1.9
itsdangerous==2.2.0
jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.8
lazy-loader==0.4
litserve==0.2.6
lmdb==1.6.2
loguru==0.7.2
lxml==5.3.1
magic-pdf==1.2.2
markdown==3.7
markupsafe==3.0.2
matplotlib==3.10.1
minio==7.1.15
modelscope==1.23.2
more-itertools==10.6.0
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.15
mypy-extensions==1.0.0
mysql-connector-python==9.2.0
networkx==3.4.2
numpy==1.26.4
omegaconf==2.3.0
onnxruntime==1.20.1
openai==1.65.4
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
opt-einsum==3.3.0
packaging==24.2
paddleocr==2.7.3
paddlepaddle==2.6.2
pandas==2.2.3
pathspec==0.12.1
pdf2docx==0.5.8
pdfminer-six==20231228
pillow==11.1.0
platformdirs==4.3.6
portalocker==3.1.1
premailer==3.10.0
prometheus-client==0.17.1
propcache==0.3.0
protobuf==6.30.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.6
pydantic-core==2.27.2
pymupdf==1.24.14
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-multipart==0.0.6
pytz==2025.1
pyyaml==6.0.1
pyzmq==26.2.1
rapid-table==1.0.3
rapidfuzz==3.12.2
rapidocr-onnxruntime==1.4.4
rapidocr-paddle==1.4.5
rarfile==4.2
redis==4.6.0
regex==2024.11.6
requests==2.32.3
robust-downloader==0.0.2
s3transfer==0.11.4
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
setuptools==75.8.2
shapely==2.0.7
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
starlette==0.27.0
stringzilla==3.12.2
struct-eqtable==0.3.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
thop==0.1.1.post2209072238
threadpoolctl==3.5.0
tifffile==2025.2.18
timm==0.9.16
tokenizers==0.19.1
tomli==2.2.1
torch==2.3.1
torchtext==0.18.0
torchvision==0.18.1
tqdm==4.67.1
transformers==4.42.4
typing-extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.85
ultralytics-thop==2.0.14
unimernet==0.2.3
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
visualdl==2.5.3
wand==0.6.13
watchfiles==1.0.4
wcwidth==0.2.13
webdataset==0.2.111
websockets==15.0.1
werkzeug==3.1.3
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@9604f5995cc628619f0e4fd913453b4d7d61db3f